+++ packages/react/src/experimental/Navigation/Header/ProductUpdates/index.tsx
@@ -15,12 +15,18 @@ import {
   DropdownMenuTrigger,
 } from "@/ui/dropdown-menu"
 import { Skeleton } from "@/ui/skeleton"
-import { ComponentProps, ReactElement, useCallback, useState } from "react"
+import {
+  ComponentProps,
+  ReactElement,
+  useCallback,
+  useEffect,
+  useState,
+} from "react"
 
 type ProductUpdate = {
   title: string
   href: string
-  imageURL: string
+  mediaUrl: string
   updated: string
   unread?: boolean
   onClick?: ComponentProps<typeof DropdownMenuItem>["onClick"]
@@ -32,6 +38,7 @@ type ProductUpdatesProp = {
   updatesPageUrl: string
   getUpdates: () => Promise<Array<ProductUpdate>>
   hasUnread?: boolean
+  currentModule: string
   onOpenChange?: ComponentProps<typeof DropdownMenu>["onOpenChange"]
   onHeaderClick?: ComponentProps<typeof DropdownMenuTrigger>["onClick"]
   onItemClick?: ComponentProps<typeof DropdownMenuItem>["onClick"]
@@ -48,6 +55,7 @@ type ProductUpdatesProp = {
 }
 
 const ProductUpdates = ({
+  currentModule,
   label,
   moreUpdatesLabel,
   getUpdates,
@@ -62,6 +70,12 @@ const ProductUpdates = ({
   const [state, setState] = useState<"idle" | "fetching" | "error">("idle")
   const [updates, setUpdates] = useState<Array<ProductUpdate> | null>(null)
   const [featuredUpdate, ...restUpdates] = updates ?? []
+
+  useEffect(() => {
+    setUpdates(null)
+    setState("idle")
+  }, [currentModule])
+
   const invokeGetUpdates = useCallback(async () => {
     try {
       setState("fetching")
@@ -149,12 +163,15 @@ const ProductUpdates = ({
 const FeaturedDropdownItem = ({
   title,
   href,
-  imageURL,
+  mediaUrl,
   unread,
   updated,
   onClick,
 }: ProductUpdate) => {
   const itemClass = "flex flex-col items-stretch gap-3 w-full"
+
+  const isVideo = mediaUrl?.includes(".mp4")
+
   return (
     <DropdownMenuItem asChild onClick={onClick}>
       <Link
@@ -167,13 +184,26 @@ const FeaturedDropdownItem = ({
           "text-f1-foreground no-underline hover:cursor-pointer"
         )}
       >
-        <div className="overflow-clip rounded border border-solid border-f1-border-secondary">
-          <Image
-            fetchPriority="high"
-            src={imageURL}
-            className="block aspect-video w-full bg-f1-background-secondary object-contain object-center"
-          />
-        </div>
+        {isVideo ? (
+          <div className="overflow-clip rounded border border-solid border-f1-border-secondary">
+            <video
+              src={mediaUrl}
+              className="block aspect-video w-full bg-f1-background-secondary object-contain object-center"
+              autoPlay
+              muted
+              loop
+              playsInline
+            />
+          </div>
+        ) : (
+          <div className="overflow-clip rounded border border-solid border-f1-border-secondary">
+            <Image
+              fetchPriority="high"
+              src={mediaUrl}
+              className="block aspect-video w-full bg-f1-background-secondary object-contain object-center"
+            />
+          </div>
+        )}
         <div className="flex items-start gap-4">
           <div className="flex-1 *:text-base">
             <h3 className="font-medium">{title}</h3>